refactor: `MenuGroup`, `Menu` 의존성 제거
