+++ src/test/java/kitchenpos/AcceptanceTest.java
@@ -8,7 +8,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.math.BigDecimal;
 import java.util.List;
-import kitchenpos.menu.dto.request.CreateMenuGroupRequest;
+import kitchenpos.menugroup.dto.request.CreateMenuGroupRequest;
 import kitchenpos.menu.dto.request.CreateMenuRequest;
 import kitchenpos.menu.dto.request.MenuProductRequest;
 import kitchenpos.order.domain.OrderStatus;
@@ -52,8 +52,10 @@ public class AcceptanceTest {
 
         Long 테이블_아이디1 = 테이블_생성();
         Long 테이블_아이디2 = 테이블_생성();
+        Long 테이블_아이디3 = 테이블_생성();
+
+        빈_테이블을_주문_테이블로_변경(테이블_아이디3);
         Long 테이블_그룹_아이디 = 테이블_그룹_생성(테이블_아이디1, 테이블_아이디2);
-        빈_테이블을_주문_테이블로_변경(테이블_아이디1);
 
         Long 주문_아이디 = 주문(메뉴_아이디, 테이블_아이디1);
         주문_테이블_손님_수_설정(테이블_아이디1);
