+++ src/test/java/kitchenpos/fixture/MenuFixture.java
@@ -2,11 +2,10 @@ package kitchenpos.fixture;
 
 import java.math.BigDecimal;
 import kitchenpos.menu.domain.Menu;
-import kitchenpos.menu.domain.MenuGroup;
 
 public class MenuFixture {
 
-    public static Menu of(String name, BigDecimal price, MenuGroup menuGroup) {
-        return new Menu(name, price, menuGroup);
+    public static Menu of(Long menuGroupId, String name, BigDecimal price) {
+        return new Menu(menuGroupId, name, price);
     }
 }
