+++ src/test/java/kitchenpos/menu/service/MenuServiceTest.java
@@ -8,11 +8,11 @@ import java.util.List;
 import kitchenpos.common.vo.PriceIsNegativeException;
 import kitchenpos.fixture.MenuGroupFixture;
 import kitchenpos.fixture.ProductFixture;
-import kitchenpos.menu.domain.MenuGroup;
+import kitchenpos.menugroup.domain.MenuGroup;
 import kitchenpos.menu.dto.request.CreateMenuRequest;
 import kitchenpos.menu.dto.request.MenuProductRequest;
 import kitchenpos.menu.dto.response.MenuResponse;
-import kitchenpos.menu.exception.MenuGroupNotFoundException;
+import kitchenpos.menugroup.exception.MenuGroupNotFoundException;
 import kitchenpos.menu.exception.MenuPriceIsBiggerThanActualPriceException;
 import kitchenpos.product.domain.Product;
 import kitchenpos.supports.ServiceTestContext;
